Moving setupbase to a filename intended to be setupbase as we got it from jupyter-project, setupbase_jp
